# swagger/public/swagger.json
{"info":{"title":"Node Swagger API","version":"1.0.0","description":"Node swagger","contacts":{"name":"Node","url":"https://github.com/G-MontaG/node","email":"arthur.osipenko@gmail.com"},"license":{"name":"MIT","url":"https://github.com/G-MontaG/node/blob/master/LICENSE"}},"host":"localhost:3000","basePath":"/","swagger":"2.0","paths":{"/api/budgets/{budgetId}":{"get":{"summary":"Get budget by ID","description":"","tags":["API"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"path","name":"budgetId","type":"string","required":true,"description":"ID of the budget to get"}],"responses":{"200":{"description":"Budget find successful","schema":{"type":"object","$ref":"#/definitions/BudgetResponse"}}}}},"/api/budgets":{"get":{"summary":"Get all budgets","description":"","tags":["API"],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"Budgets find successful","schema":{"type":"array","items":{"$ref":"#/definitions/BudgetResponse"}}}}}},"/api/user":{"get":{"summary":"Get user by ID","description":"","tags":["API"],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"User find successful","schema":{"type":"object","$ref":"#/definitions/UserResponse"}}}},"put":{"summary":"User profile update","description":"","tags":["API"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Update info","required":true,"schema":{"$ref":"#/definitions/UserUpdate"}}],"responses":{"200":{"description":"User update successful","schema":{"type":"object","$ref":"#/definitions/UserResponse"}}}}},"/auth/forgot/email":{"post":{"summary":"Forgot password, send email token to verify user existing","description":"","tags":["Auth"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"","required":true,"schema":{"$ref":"#/definitions/ForgotEmail"}}],"responses":{"200":{"description":"Verify email and send token successful","schema":{"type":"object","$ref":"#/definitions/ForgotEmailResponse"}}}}},"/auth/forgot/token":{"post":{"summary":"Forgot password, verify token from email and set new password","description":"","tags":["Auth"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"","required":true,"schema":{"$ref":"#/definitions/ForgotToken"}}],"responses":{"200":{"description":"Verify token and set new password successful","schema":{"type":"object","$ref":"#/definitions/ForgotTokenResponse"}}}}},"/auth/login":{"post":{"summary":"Login to the application","description":"","tags":["Auth"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"","required":true,"schema":{"$ref":"#/definitions/Login"}}],"responses":{"200":{"description":"Login successful","schema":{"type":"object","$ref":"#/definitions/AuthTokenResponse"}}}}},"/auth/reset/email":{"post":{"summary":"Reset password, send email token to verify user existing","description":"","tags":["Auth"],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"Send email token successful","schema":{"type":"object","$ref":"#/definitions/ResetEmailResponse"}}}}},"/auth/reset/token":{"post":{"summary":"Reset password, verify token from email and set new password","description":"","tags":["Auth"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"","required":true,"schema":{"$ref":"#/definitions/ResetToken"}}],"responses":{"200":{"description":"Verify token and set new password successful","schema":{"type":"object","$ref":"#/definitions/ResetTokenResponse"}}}}},"/auth/sign-up":{"post":{"summary":"Sign-up to the application","description":"","tags":["Auth"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"","required":true,"schema":{"$ref":"#/definitions/SignUp"}}],"responses":{"200":{"description":"Sign-up successful","schema":{"type":"object","$ref":"#/definitions/AuthTokenResponse"}}}}},"/auth/verify-email":{"post":{"summary":"Verify user email","description":"","tags":["Auth"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"","required":true,"schema":{"$ref":"#/definitions/VerifyEmail"}}],"responses":{"200":{"description":"Verify email successful","schema":{"type":"object","$ref":"#/definitions/VerifyEmailResponse"}}},"security":[{"Authorization":[]}]}}},"definitions":{"BudgetResponse":{"type":"object","properties":{"name":{"type":"string"},"currency":{"type":"boolean"},"numberFormat":{"type":"string"},"currencyPlacement":{"type":"string"},"dateFormat":{"type":"string"}},"required":["name","currency","numberFormat","currencyPlacement","dateFormat"]},"UserResponse":{"type":"object","properties":{"email":{"type":"string"},"emailConfirmed":{"type":"boolean"},"firstName":{"type":"string"},"lastName":{"type":"string"},"gender":{"type":"string"},"language":{"type":"string"}},"required":["email","emailConfirmed","firstName","lastName","gender","language"]},"UserUpdate":{"type":"object","properties":{"firstName":{"type":"string","minLength":2,"maxLength":100},"lastName":{"type":"string","minLength":2,"maxLength":100},"gender":{"type":"string"},"language":{"type":"string"}}},"AuthTokenResponse":{"required":["message","token"],"properties":{"message":{"type":"string"},"token":{"type":"string"}}},"ForgotEmail":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ForgotEmailResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"ForgotToken":{"type":"object","properties":{"token":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","minLength":3,"maxLength":30},"confirm_password":{"type":"string","minLength":3,"maxLength":30}},"required":["token","email","password","confirm_password"]},"ForgotTokenResponse":{"type":"object","properties":{"message":{"type":"string"},"token":{"type":"string"}},"required":["message","token"]},"Login":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":3,"maxLength":30}},"required":["email","password"]},"ResetEmailResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"ResetToken":{"type":"object","properties":{"token":{"type":"string"},"old_password":{"type":"string","minLength":3,"maxLength":30},"password":{"type":"string","minLength":3,"maxLength":30},"confirm_password":{"type":"string","minLength":3,"maxLength":30}},"required":["token","old_password","password","confirm_password"]},"ResetTokenResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"SignUp":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":3,"maxLength":30}},"required":["email","password"]},"VerifyEmail":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"VerifyEmailResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}},"responses":{},"parameters":{},"securityDefinitions":{"Authorization":{"type":"apiKey","name":"Authorization","in":"header"}},"tags":[{"name":"API","description":"Application API"},{"name":"Auth","description":"Application authorization"}]}